test-requirements.txt
nose
mock
django_nose
